% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results.R
\name{anova_results}
\alias{anova_results}
\title{Results of ANOVA in APA standard}
\usage{
anova_results(y, x)
}
\arguments{
\item{y}{a numeric vector (dependent variable)}

\item{x}{a factor (grouping variable)}
}
\value{
"F(df, df) = F; p = p"
}
\description{
Function to print results of an ANOVA in APA standard (F(df, df) = F; p = p)
}
\examples{
set.seed(1)
df <- data.frame(happy = c(rnorm(n = 10, mean = 3, sd = 0.5),
                           rnorm(n = 10, mean = 3.4, sd = 0.5),
                           rnorm(n = 10, mean = 3, sd = 0.5)),
                 treatment = factor(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                                      1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
                                      2, 2, 2, 2, 2, 2, 2, 2, 2, 2)))

anova_results(df$happy, df$treatment)

}
\author{
Dominik Vogel
}
