% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results.R
\name{chi_results}
\alias{chi_results}
\title{Results of Chi2-test in APA standard}
\usage{
chi_results(x, y)
}
\arguments{
\item{x}{a numeric vector or a factor}

\item{y}{a numeric vector or a factor}
}
\value{
"Chi2(df) = chi2; p = p"
}
\description{
Function to print results of a Chi2-test in APA standard (Chi2(df) = chi2; p = p)
}
\examples{
df <- data.frame(id = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
                  treatment = factor(c("treated", "treated", "not-treated", "treated",
                                       "treated", "treated", "not-treated",
                                       "treated", "not-treated", "treated")),
                   improvement = factor(c("improved","improved","not-improved","improved",
                                          "improved", "improved", "not-improved",
                                          "not-improved", "not-improved", "improved")))
chi_results(df$treatment, df$improvement)

}
\author{
Dominik Vogel
}
