% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{corstars}
\alias{corstars}
\title{Title}
\usage{
corstars(x, method = c("pearson", "spearman"), removeTriangle = c("upper",
  "lower"), result = c("none", "html", "latex"), labels_rows,
  labels_cols = 1:length(labels_rows) - 1, sig.level = 0.05,
  caption = c("Correlation"), filename = "")
}
\arguments{
\item{x}{a matrix containing the data}

\item{method}{correlation method. "pearson"" or "spearman"" is supported}

\item{removeTriangle}{remove upper or lower triangle}

\item{result}{Print result in Console ("none"), generate HTML file ("html"), generate latex file ("latex")}

\item{labels_rows}{Labels for the rows (i.e., variable names). Length musst be same as number of variables}

\item{labels_cols}{Labels for columns. Length musst be same as number of variables - 1}

\item{sig.level}{Significance level (.1 or .05)}

\item{caption}{Caption for the table}

\item{filename}{File name to save output to}
}
\value{
Correlation table in console or file
}
\description{
Title
}
\examples{
# Console output
corstars(mtcars, method="pearson", removeTriangle="upper", result="none",
   caption = "Correlations",
   sig.level = 0.1,
   labels_rows = c("(1) mpg", "(2) cyl", "(3) disp", "(4) hp",
                   "(5) drat", "(6) wt", "(7) qsec", "(8) vs",
                   "(9) am", "(10) gear",
                   "(11) carb"),
   labels_cols = 1:10)

# HTML output
corstars(mtcars, method="pearson", removeTriangle="upper", result="html",
   caption = "Correlations", filename = "corr.html",
   sig.level = 0.1,
   labels_rows = c("(1) mpg", "(2) cyl", "(3) disp", "(4) hp",
                   "(5) drat", "(6) wt", "(7) qsec", "(8) vs",
                   "(9) am", "(10) gear",
                   "(11) carb"),
   labels_cols = 1:10)

}
\author{
Dominik Vogel
}
