% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_alpha}
\alias{extract_alpha}
\title{Extract Cronbach's Alpha}
\usage{
extract_alpha(vars, data)
}
\arguments{
\item{vars}{Variables to include in the analysis}

\item{data}{Data frame to select variables from}
}
\value{
Alpha value
}
\description{
Function to perform a Cronbach's Alpha analysis and
to extract the Alpha value
}
\examples{
extract_alpha(vars = c("cyl", "hp", "wt"), data = mtcars)

}
\author{
Dominik Vogel
}
