% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_hlm}
\alias{extract_hlm}
\title{Extract parameters from nlme::lme-models}
\usage{
extract_hlm(model, var, parameter)
}
\arguments{
\item{model}{The nlme::lme-model (must be build in advance)}

\item{var}{Name of Variable for with parameter should be extracted. Put in "".}

\item{parameter}{"b" = coefficient, "p" = p-value}
}
\value{
"b = ..." or "p = ..."
}
\description{
Function extracts coefficients and p-values from nlme::lme-models
}
\examples{
model <- nlme::lme(mpg ~ cyl, random = ~1|am, data = mtcars)
extract_hlm(model, "cyl", "b")

}
\author{
Dominik Vogel
}
