% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_lm}
\alias{extract_lm}
\title{Extract parameters from stats::lm-models}
\usage{
extract_lm(model, var, parameter)
}
\arguments{
\item{model}{The stats::lm-model (must be build in advance)}

\item{var}{Name of Variable for with parameter should be extracted. Put in "".}

\item{parameter}{"b" = coefficient, "p" = p-value}
}
\value{
returns "b = ..." or "p = ..."
}
\description{
Function extracts coefficients and p-values from stats::lm-models
}
\examples{
model <- stats::lm(mpg ~ cyl + am, data = mtcars)
extract_lm(model, "cyl", "p")

}
\author{
Dominik Vogel
}
