% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey.R
\name{lime_names}
\alias{lime_names}
\title{Fix variable names for data frames exportet from LimeSurvey}
\usage{
lime_names(df)
}
\arguments{
\item{df}{Data frame from LimeSurvey}
}
\value{
df
}
\description{
Function to get rid of the 'X[Y]' variable names in data frames exported from LimeSurvey
}
\examples{
lime_names(mtcars)

}
\author{
Dominik Vogel
}
