% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.R
\name{mean_index2}
\alias{mean_index2}
\title{Add mean index to data frame}
\usage{
mean_index2(df, name, vars)
}
\arguments{
\item{df}{Data frame to add index to}

\item{name}{Name of the new index}

\item{vars}{Variables to combine to an index}
}
\value{
df
}
\description{
Add a mean index of a defined group of variables to a data frame
}
\examples{
df <- mtcars
mean_index2(df, "mean", vars = c(c("gear", "carb")))
}
\author{
Dominik Vogel
}
