% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myskim.R
\name{myskim}
\alias{myskim}
\title{Print discriptives using skimr::skim}
\usage{
myskim(df)
}
\arguments{
\item{df}{Data frame to get descriptives from}
}
\description{
Function to effortless use skimr::skim() in Windows.
Problem: histogram functionality of skimr::skim works only in Windows if locale is set to Chinese
Solution: Set locale to Chinese and back to German afterwards
}
\examples{
myskim(mtcars)

}
\seealso{
skimr::skim()
}
\author{
Dominik Vogel
}
