% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{t_test2}
\alias{t_test2}
\title{t-test from means and sds}
\usage{
t_test2(m1, m2, s1, s2, n1, n2, m0 = 0, equal.variance = FALSE)
}
\arguments{
\item{m1}{Sample mean group 1}

\item{m2}{Sample mean group 2}

\item{s1}{Sample S.D. group 1}

\item{s2}{Sample S.D. group 2}

\item{n1}{Sample size group 1}

\item{n2}{Sample size group 1}

\item{m0}{the null value for the difference in means to be tested for. Default is 0.}

\item{equal.variance}{Whether or not to assume equal variance. Default is FALSE.}
}
\value{
returns vector with difference of means, standard error, t-value, p-value
}
\description{
Function to perform a two independent sample t-test
based on sample size, means, and standard deviations
}
\note{
Copied from https://stats.stackexchange.com/questions/30394/how-to-perform-two-sample-t-tests-in-r-by-inputting-sample-statistics-rather-tha
}
\examples{
t <- t_test2(3.5, 4, 0.7, 0.8, 100, 200)

}
\references{
https://stats.stackexchange.com/questions/30394/how-to-perform-two-sample-t-tests-in-r-by-inputting-sample-statistics-rather-tha
}
\author{
Dominik Vogel
}
